split simulation code from image utils
simulation code was not all in one place.
